pyproject.toml
[project]
name = "confy"
version = "1.0.5"
description = "Aplicativo desktop para o sistema Confy de comunicação criptografada."
authors = [
    {name = "Henrique Sebastião", email = "contato@henriquesebastiao.com"},
    {name = "Valério Gonçalves", email = "valeriogoncalvesdev@gmail.com"},
    {name = "Lucas Sousa", email = "lucasdedeus-15@hotmail.com"},
    {name = "Wanderson Amorim", email = "wandersonfla100@hotmail.com"},
    {name = "Emanuel Chagas", email = "pinheiroemanuel192@gmail.com"},
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = "==3.13.7"
dependencies = [
    "pyside6-essentials (>=6.9.1,<7.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "confy-addons (>=1.1.0,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.12"
taskipy = "^1.14.1"
radon = "^6.0.1"
bandit = "^1.8.6"
pytest = "^8.4.2"
pytest-qt = "^4.5.0"
pytest-cov = "^6.3.0"
briefcase = "^0.3.25"

[tool.ruff]
line-length = 99

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'D', 'UP', 'PERF']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.ruff.lint.per-file-ignores]
'qss.py' = ['W293']
'connect_to_user.py' = ['W293']
'chat.py' = ['W293', 'E501', 'PLR0912', 'PLR0915']
'__init__.py' = ['F401', 'D104']
'tests/**' = ['D']
'utils.py' = ['D105', 'D101']
'*.py' = ['D401']

[tool.taskipy.tasks]
lint = 'poetry run ruff check .; poetry run ruff check . --diff'
format = 'poetry run ruff format .; poetry run ruff check . --fix'
radon = 'poetry run radon cc ./confy -a -na'
bandit = 'poetry run bandit -r ./confy'
changelog = 'github_changelog_generator -u confy-security -p app -o CHANGELOG --no-verbose'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
qt_api = 'pyside6'

[tool.coverage.run]
branch = true
omit = ["__main__.py", "utils.py"]

[tool.briefcase]
project_name = "Confy"
version = "1.0.5"
author = "Confy Security"
author_email = "confy@henriquesebastiao.com"
url = "https://github.com/confy-security/app"
bundle = "com.henriquesebastiao"
license.file = "LICENSE"

[tool.briefcase.app.confy]
sources = ["confy"]
test_sources = ["tests"]
description = "Confy Desktop application."
long_description = "Desktop application for the Confy encrypted communication system."
icon = "confy/assets/icon"
installer_icon = "confy/assets/icon"
formal_name = "Confy"

requires = [
    "pyside6-essentials (>=6.9.1,<7.0.0)",
]

test_requires = [
    "pytest",
    "pytest-qt",
]

[tool.briefcase.app.confy.macOS]
universal_build = true
requires = [
]

[tool.briefcase.app.confy.linux]
requires = [
]

[tool.briefcase.app.confy.linux.system.debian]
system_requires = [
]

system_runtime_requires = [
]

[tool.briefcase.app.confy.linux.system.rhel]
system_requires = [
]

system_runtime_requires = [
]

[tool.briefcase.app.confy.linux.system.suse]
system_requires = [
]

system_runtime_requires = [
]

[tool.briefcase.app.confy.linux.system.arch]
system_requires = [
]

system_runtime_requires = [
]

[tool.briefcase.app.confy.linux.appimage]
manylinux = "manylinux_2_28"

system_requires = [
]

linuxdeploy_plugins = [
]

[tool.briefcase.app.confy.linux.flatpak]
flatpak_runtime = "org.freedesktop.Platform"
flatpak_runtime_version = "24.08"
flatpak_sdk = "org.freedesktop.Sdk"

[tool.briefcase.app.confy.windows]
requires = [
]

# Mobile deployments
[tool.briefcase.app.confy.iOS]
supported = false

[tool.briefcase.app.confy.android]
supported = false

# Web deployments
[tool.briefcase.app.confy.web]
supported = false